TextBuilderDev

Documentation

data TextBuilder

Specification of how to efficiently construct strict Text. Provides instances of Semigroup and Monoid, which have complexity of O(1).

Accessors

buildText :: TextBuilder -> Text

Execute a builder producing a strict text.

length :: TextBuilder -> Int

Get the amount of characters.

null :: TextBuilder -> Bool

Check whether the builder is empty.

Output IO

putToStdOut :: TextBuilder -> IO ()

Put builder, to stdout.

putToStdErr :: TextBuilder -> IO ()

Put builder, to stderr.

putLnToStdOut :: TextBuilder -> IO ()

Put builder, followed by a line, to stdout.

putLnToStdErr :: TextBuilder -> IO ()

Put builder, followed by a line, to stderr.

Constructors

Builder manipulators

force :: TextBuilder -> TextBuilder

Run the builder and pack the produced text into a new builder.

Useful to have around builders that you reuse, because a forced builder is much faster, since it's virtually a single call memcopy.

intercalate :: Foldable f => TextBuilder -> f TextBuilder -> TextBuilder

Intercalate builders.

intercalateMap :: Foldable f => TextBuilder -> (a -> TextBuilder) -> f a -> TextBuilder

Intercalate projecting values to builder.

padFromLeft :: Int -> Char -> TextBuilder -> TextBuilder

Pad a builder from the left side to the specified length with the specified character.

padFromRight :: Int -> Char -> TextBuilder -> TextBuilder

Pad a builder from the right side to the specified length with the specified character.

Textual

text :: Text -> TextBuilder

Strict text.

lazyText :: Text -> TextBuilder

Lazy text.

string :: String -> TextBuilder

String.

asciiByteString :: ByteString -> TextBuilder

ASCII byte string.

It's your responsibility to ensure that the bytes are in proper range, otherwise the produced text will be broken.

hexData :: ByteString -> TextBuilder

Hexadecimal readable representation of binary data.

Character

char :: Char -> TextBuilder

Unicode character.

Low-level character

unicodeCodePoint :: Int -> TextBuilder

Unicode code point.

utf16CodeUnits1 :: Word16 -> TextBuilder

Single code-unit UTF-16 character.

utf16CodeUnits2 :: Word16 -> Word16 -> TextBuilder

Double code-unit UTF-16 character.

utf8CodeUnits1 :: Word8 -> TextBuilder

Single code-unit UTF-8 character.

utf8CodeUnits2 :: Word8 -> Word8 -> TextBuilder

Double code-unit UTF-8 character.

utf8CodeUnits3 :: Word8 -> Word8 -> Word8 -> TextBuilder

Triple code-unit UTF-8 character.

utf8CodeUnits4 :: Word8 -> Word8 -> Word8 -> Word8 -> TextBuilder

UTF-8 character out of 4 code units.

Integers

Decimal

decimal :: Integral a => a -> TextBuilder

Decimal representation of an integral value.

unsignedDecimal :: Integral a => a -> TextBuilder

Decimal representation of an unsigned integral value.

fixedUnsignedDecimal :: Integral a => Int -> a -> TextBuilder

thousandSeparatedDecimal :: Integral a => Char -> a -> TextBuilder

Decimal representation of an integral value with thousands separated by the specified character.

thousandSeparatedUnsignedDecimal :: Integral a => Char -> a -> TextBuilder

Decimal representation of an unsigned integral value with thousands separated by the specified character.

dataSizeInBytesInDecimal :: Integral a => Char -> a -> TextBuilder

Data size in decimal notation over amount of bytes.

Binary

unsignedBinary :: Integral a => a -> TextBuilder

Unsigned binary number.

unsignedPaddedBinary :: (Integral a, FiniteBits a) => a -> TextBuilder

Unsigned binary number.

finiteBitsUnsignedBinary :: FiniteBits a => a -> TextBuilder

A less general but faster alternative to unsignedBinary.

Hexadecimal

hexadecimal :: Integral a => a -> TextBuilder

Hexadecimal representation of an integral value.

unsignedHexadecimal :: Integral a => a -> TextBuilder

Unsigned hexadecimal representation of an integral value.

Digits

decimalDigit :: Integral a => a -> TextBuilder

Decimal digit.

hexadecimalDigit :: Integral a => a -> TextBuilder

Hexadecimal digit.

Real

fixedDouble

Arguments

:: Int
Amount of decimals after point.

-> Double

-> TextBuilder

Double with a fixed number of decimal places.

doublePercent

Arguments

:: Int
Amount of decimals after point.

-> Double

-> TextBuilder

Double multiplied by 100 with a fixed number of decimal places applied and followed by a percent-sign.

Time

utcTimeInIso8601 :: UTCTime -> TextBuilder

utcTimestampInIso8601

Arguments

:: Int
Year.

-> Int
Month.

-> Int
Day.

-> Int
Hour.

-> Int
Minute.

-> Int
Second.

-> TextBuilder

General template for formatting date values according to the ISO8601 standard. The format is the following:

2021-11-24T12:11:02Z

Integrations with various time-libraries can be easily derived from that.

intervalInSeconds :: RealFrac seconds => seconds -> TextBuilder

Time interval in seconds. Directly applicable to DiffTime and NominalDiffTime.

Classes

class IsomorphicToTextBuilder a where

Evidence that there exists an unambiguous way to convert a type to and from TextBuilder.

Unlike conversion classes from other libs this class is lawful. The law is:

fromTextBuilder . toTextBuilder = id

This class does not provide implicit rendering, such as from integer to its decimal representation. There are multiple ways of representing an integer as text (e.g., hexadecimal, binary). The non-ambiguity is further enforced by the presence of the inverse conversion. In the integer case there is no way to read it from a textual form without a possibility of failing (e.g., when the input string cannot be parsed as an integer).

If you're looking for such conversion classes, this library is not a place for them, since there can be infinite amount of flavours of conversions. They are context-dependent and as such should be defined as part of the domain.

Methods

toTextBuilder :: a -> TextBuilder

fromTextBuilder :: TextBuilder -> a